+++ packages/terra-status-view/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.55.0 - (April 27, 2023)
+
 * Added
   * Added dynamic example for terra-status-view.